[Agent] refactor placeholder and LLM utils
